Replace writeByte with writeShort in TLSUtilTests (#933)
(cherry picked from commit 059b11d87c9084cce2dbc97a5f90db25b6e39cef)

+++ src/test/java/com/amazon/opendistroforelasticsearch/security/ssl/util/TLSUtilTests.java
@@ -72,7 +72,7 @@ public class TLSUtilTests {
             buffer.writeByte(TLS_MAJOR_VERSION);
             buffer.writeByte(TLS_MINOR_VERSION);
             //setting content length as 0
-            buffer.writeByte(0);
+            buffer.writeShort(0);
             Assert.assertFalse(TLSUtil.isTLS(buffer));
         }
